+++ cvat/apps/git/git.py
@@ -10,7 +10,7 @@ from cvat.apps.engine.models import Task, Job, User
 from cvat.apps.engine.annotation import dump_task_data
 from cvat.apps.engine.plugins import add_plugin
 from cvat.apps.git.models import GitStatusChoice
-from cvat.apps.annotation.models import AnnotationHandler
+from cvat.apps.annotation.models import AnnotationDumper
 
 from cvat.apps.git.models import GitData
 from collections import OrderedDict
@@ -264,7 +264,7 @@ class Git:
         timestamp = datetime.datetime.now().strftime("%Y_%m_%d_%H_%M_%S")
         display_name = "CVAT XML 1.1"
         display_name += " for images" if self._task_mode == "annotation" else " for videos"
-        cvat_dumper = AnnotationHandler.objects.get(display_name=display_name)
+        cvat_dumper = AnnotationDumper.objects.get(display_name=display_name)
         dump_name = os.path.join(db_task.get_task_dirname(),
             "git_annotation_{}.".format(timestamp) + "dump")
         dump_task_data(